Only reset lastScrollTime on non-timed-out scroll
This change addresses an issue where the slides appear unresponsive if a
user swipes or scrolls multiple times rapidly within a given period.

The behavior was to simply not scroll till the user stopped attempting
to change slides and paused. This can induce some frustration in users.
Instead this will allow users to scroll once the timeout is up.